+++ Runtime/Loggers/InternalLoggerBase.cs
@@ -4,6 +4,14 @@ namespace DTech.Logging
 {
 	internal abstract class InternalLoggerBase : ILogger
 	{
+		protected string Tag { get; }
+		protected string NullStateName => nameof(NullState);
+
+		public InternalLoggerBase(string tag)
+		{
+			Tag = tag;
+		}
+		
 		public IDisposable BeginScope<TState>()
 		{
 			return BeginScope(nameof(TState));